src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/activeelement-should-be-shadowhost-expected.txt

CONSOLE ERROR: line 32: Uncaught TypeError: shadowHost.createShadowRoot is not a function
Makes sure that document.activeElement returns a shadow host when a element in the correspoinding shadow tree is focused.
